+++ main.js
@@ -1,8 +1,8 @@
 import './style.css'
 import javascriptLogo from './javascript.svg'
 import viteLogo from '/vite.svg'
-import {setupCounter} from './counter.js'
-import {setupThree} from "./three.js";
+import {setupCounter} from './src/counter.js'
+import {setupThree} from "./src/three.js";
 
 document.querySelector('#app').innerHTML = `
   <div>
